pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "A difference calculator is a program that determines the difference between two data structures."
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.10"
]
authors = ["Rust <afarg1710@gmail.com>"]
repository = "https://github.com/RustSaf/python-project-50"
readme = "README.md"
packages = [{ include = "gendiff" }, { include = "tests" },]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^5.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"
pytest-cov = "^5.0.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
